% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidecounter_count.R
\name{guidecounter_count}
\alias{guidecounter_count}
\title{Run guide-counter count via the internal Rust wrapper}
\usage{
guidecounter_count(input, library, offset_min_fraction, output, exact_match)
}
\arguments{
\item{input}{A character vector of input file paths.}

\item{library}{A string specifying the library name or path to use.}

\item{offset_min_fraction}{A numeric value specifying the minimum offset threshold.}

\item{output}{A string specifying the output file path.}
}
\description{
Run guide-counter count via the internal Rust wrapper
}
