% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{guidecounter_count_internal}
\alias{guidecounter_count_internal}
\title{Exposes the guidecounter_count_internal functionality as a Rust function callable from R.}
\usage{
guidecounter_count_internal(
  input,
  library,
  offset_min_fraction,
  output,
  exact_match,
  verbose
)
}
\arguments{
\item{input}{A character vector of input file paths.}

\item{library}{A string specifying the library name or path to use.}

\item{offset_min_fraction}{A numeric value specifying the minimum offset threshold.}

\item{output}{A string specifying the output file path.}

\item{verbose}{Logical; emit progress messages from guide-counter.}
}
\value{
A character string indicating success or error message.
}
\description{
Exposes the guidecounter_count_internal functionality as a Rust function callable from R.
}
