% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{guidecounter_count}
\alias{guidecounter_count}
\title{Function to call the fqtk CLI command for demux.
Exposes the demux functionality as a Rust function that can be called from R.}
\usage{
guidecounter_count(input, library, offset_min_fraction, output)
}
\description{
Function to call the fqtk CLI command for demux.
Exposes the demux functionality as a Rust function that can be called from R.
}
